refactor(MenuValidator): MenuService -> ProductRepository 의존성 제거
