+++ todo/step3.md
@@ -77,7 +77,7 @@ Lombok은 그 강력한 기능만큼 사용상 주의를 요한다.
 - [x] 모델링 분석
 ![step3_image5_kitchenpos.jpg](images/step3_image5_kitchenpos.jpg)
 - [x] OrderTable - TableGroup 양방향 도메인 의존성 제거
-- [ ] Menu - Product 패키지 의존성 제거
+- [x] Menu - Product 패키지 의존성 제거
 - [ ] Order - Table 패키지 의존성 제거
 - [ ] Menu - Order 패키지 의존성 제거
 - [ ] 기존 리뷰 반영 (Message 개선 - static 클래스 -> Enum)
